backend/ocr-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
paddleocr==2.9.1
paddlepaddle==3.0.0
opencv-python-headless==4.10.0.84
Pillow==11.0.0
pydantic==2.9.2
pydantic-settings==2.6.1
